Reorganize jobframework package.
